requirements.txt
flask
opencv-python
ultralytics
numpy
twilio
python-dotenv
